+++ CHANGELOG.md
@@ -44,6 +44,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
  - #2985, Fix not adding `application_name` on all connection strings - @steve-chavez
  - #3424, Admin `/live` and `/ready` now differentiates a failure as 500 status - @steve-chavez
     + 503 status is still given when postgREST is in a recovering state
+ - #3478, Media Types are parsed case insensitively - @develop7
 
 ### Deprecated